Rate-limit heartbeats sent because of starting a new packet
A 10k/s rate is still a bit silly; the immediate reason for
rate-limiting them is the load they put on the HEARTBEAT/ACKNACK
processing when there are new readers that haven't completed the
handshake yet, and that really only becomes a problem at the very high
packet rates that you can get on loopback interfaces.

Some refinement is probably wise.

Signed-off-by: Erik Boasson <eb@ilities.com>