Out-of-sync reader may be expecting seqno > pwr
This removes an incorrect assert checking that the proxy-writer is never
behind an out-of-sync reader in the sequence number it expects to
deliver next:

- the reader may simply be catching up after having been created, in
  which its next expected sequence number is < the next expected
  sequence of the proxy writer;

- the reader may have been created while the writer is furiously writing
  data and actually be expecting a sequence number > what the proxy
  writer is expected (for its in-sync readers).

The delivery paths for the "in sync" path always skip all out-of-sync
readers, so the second case is handled correctly.

What needs further study is under which circumstances the swtich to "in
sync" might be delayed because the switch happens when the sequence
numbers are equal but the check isn't necessarily performed for each
delivered sample.  But that is a different matter.

Signed-off-by: Erik Boasson <eb@ilities.com>